fix: error code from uint8 to int32 in error class

+++ inc/osal/error.hpp
@@ -110,7 +110,7 @@ public:
 
 private:
     char msg[128]{0};               ///< The error message.
-    uint8_t code = 0;               ///< The error code.
+    int32_t code = 0;               ///< The error code.
     char file[64]{0};               ///< The file name where the error occurred.
     char func[64]{0};               ///< The function name where the error occurred.
     uint32_t line = 0;              ///< The line number where the error occurred.
